requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
elasticsearch==7.7.1
idna==2.9
joblib==0.15.1
kafka-python==2.0.1
kazoo==2.5.0
neo4j==1.7.6
neobolt==1.7.17
neotime==1.7.4
nltk==3.5
oauthlib==3.1.0
prompt-toolkit==2.0.10
psutil==5.7.0
py2neo==4.3.0
py4j==0.10.7
Pygments==2.3.1
PySocks==1.7.1
pyspark==2.2.3
pytz==2020.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
setuptools-scm==4.1.2
six==1.15.0
tabulate==0.8.7
textblob==0.15.3
tqdm==4.46.1
tweepy==3.8.0
urllib3==1.24.3
wcwidth==0.2.4
